fix : list 화
body : .

+++ src/main/java/com/example/tooktook/model/dto/questionDto/QuestionDto.java
@@ -8,5 +8,5 @@ import java.util.List;
 public class QuestionDto {
     private Long qid;
     private String questions;
-    private List<Integer> answerIds;
+    private List<Long> answerIds;
 }
